requirements.txt
discord.py>=2.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
httpx>=0.27.0
